Introduce general purpose menu prompt. (#90)
* Implement general method-based prompts.

Introduces MenuPrompt which takes a context object and prompt data. If a
method argument is passed with the button, it will execute that method
on the context object when clicked.

* Prevent arbitrary method calls on menu prompt.

To prevent execution of arbitrary methods on the context object, the
menu prompt should only call the method if there is a button that
corresponds to that method.