Allow dupes for unique attachments (#97)
* Detect duplicates of attached cards.
* Clean up additional find card methods.
Removes the unused findCardInPlayByCode and converts findCardByUuid and
findCardByName to use the generalized findCard method.
* Play attachments as duplicates (except setup)
* Get tests around existing Player.removeAttachment
* Handle removing an attachment w/ duplicates.
* Swap order of revealing setup cards and attaching
Per the rules, both players reveal their cards simultaneously and *then*
choose where their attachments should go. By doing it in this order, it
also ensures that duplicate unique attachments are grouped together
prior to prompting the user to attach them.
* Conditionally disable saving attachments via dupes
Dragging and dropping a card with attachments should result in all
duplicates of those attachments being discarded as well. Similarly,
there may be cards in the future that discard attachments that cannot be
saved (as there are currently for killing or sacrificing characters)
* Handle dupe attachment drag-and-drop from discard