Fix attachment overlap and menus under cards. (#150)
* Convert z-indexes into constants.
* Remove special casing for draw deck popup
* Get card layout working with flexbox
* Create separate class for card image.
Card images were previously using the `card` class. This added
properties such as box-shadow that weren't necessarily appropriate for
the image. It has been also reworked to have orientation classes to
handle dimensions and kneeling.
* Remove `.horizontal-card` for `.card.horizontal`
* Get rid of `.plot-card`
* Remove commented out styling
* Move passed in style to the card wrapper.
The overflow property for `.card` is properly set to hidden, but the
menu was being rendered within it. To avoid z-index problems, the menu
has been moved to be a sibling of the card.
* Menus should appear above pile headers.
* Handle styling for Arya Stark facedown duplicates.
* Simplify rendering of wrapped cards.
The 'selected plot' source no longer existed, so there was no reason to
modify the wrapper class anymore.