Pass cards to play-related methods and events instead of IDs (#154)
* Move active plot can play check into player.

Previously whether a card was allowed to be played by the active plot
was done in Game.playCard. The opponent's plot was also checked.
However, there are currently no cards that can prevent an opponent from
playing certain kinds of cards. If any such cards are introduced in the
future, all of the opponents cards should be asked, not just their
active plot.

* Update Player.playCard to take a card directly.

Previously, the playCard method took a card ID and optional source list
to look up the card object. With the introduction of moveCard, it is no
longer necessary to pass around the source list. Thus it's better if
playCard doesn't need to look up the card each time when it usually has
been looked up already.

* Send full card object with onCardPlayed event.

* Simplify Player.canPlay

* Don't stub card finding methods in canPlay test