Only play cards as ambush during challenge phase (#181)
Previously, if a card with the Ambush keyword was played normally during
the setup phase, it would play it as an ambushed card at its non-ambush
price, including a message saying that they ambushed it. This would
reveal information about the player's setup to the opponent before the
cards were revealed.