Fix Nymeria Sand (#223)
Previously, when the player used Nymeria's ability it would not remove
the icon from the opponents card. This is because the effect wasn't
specifically targetting opponent controlled cards. This change makes it
target a specific card, which skips the controller checks.

Additionally, Nymeria should now be able to remove an icon even if the
target card does not have that icon. This should be more in line with
the card text.