Do not declare a winner if cannot be won. (#250)
Previously, if the "cannot win a challenge" effect was in play, the
challenge would still declare a winner but skip over applying claim. Now
when the effect is in play, the attacking player is not named the
winner. If the defending player wins, they are still declared the
winner.

Fixes #247.