Fix Summer Harvest from granting too much gold (#398)
Previously, if you went through all of your plots and played Summer
Harvest a second time, it would retain its income value from the first
time in addition to how much it would normally gain this round. Now
instead of setting a gold modifier, it sets the base income value which
is overwritten each time. This prevents the previous bug and should
allow interactions with Varys' Riddle to be correct.