Only fetch images if one exists. (#425)
Previously, if the card.imagesrc property returned by ThronesDB was
null to indicate they don't have an image for the card, the fetch data
script would create a zero byte file locally. This prevented the image
from being fetched after the image had been uploaded, because there was
a local file (even though it was empty).