implement most of the missing challenge events (#415)
* refactor "challenge event" logic from Ours is the Fury to separate class

The new class is called ChallengeEvent and by default checks whether we're
currently in a challenge or not. Additionally it can be passed arguments to
declaratively request checking that we're in a challenge of a given type and/or
participating in a given role (attacker/defender).

* implement Bless Him With Salt

* implement For the North!

* implement In the Name of Your King!

* implement The Watcher on the Walls

* implement Winter Is Coming

* new effect to set the max number of challenges of a given type

This is different from increasing/decreasing the max, which was already
possible. The main use case for setting to some arbitrary value is setting it
to 0 (e.g., for In The Name of Your King), but the more general version might
be useful in the future.

* implement There Is My Claim