Only recalc targets for conditional effects. (#460)
Previously, if an effect was state dependent (dynamic strength, kill on
0 strength, etc) OR an effect was conditional, the recalculation
mechanism would always remove all targets and go through all cards for
eligible targets. This lead to a bug with Dracarys which is state
dependent, not conditional, and not persistent. The effect of Dracarys
would be re-applied if the card killed by its effect was manually
dragged back into play - the target was removed when it left play, but
the matching function would allow it to be reapplied after entering play
again.

Now, only conditional effects will update their target lists upon
recalculation. Effects that are solely state dependent will unapply and
reapply their effect to existing targets, but will not add any
additional targets.

Fixes #454.