Remove gold check from Lady In Waiting (#560)
This should make the temporary implementation work in more cases. Only case in which this won't be perfect is if you want to play her as a dupe while having less than 2 gold. Manually adding gold to be able to marshal her, then substracting the gold you got from the handler should solve that.