Hide users that have left in lobby listings. (#604)
Previously, if a player left the game after it had been started, they
could still be considered active in the game from the perspective of the
lobby. This would leave their name listed in the game and could also
potentially cause reconnect issues. Now, once they leave an active game,
they are marked as 'left' in the lobby listing.