Allow in-hand actions and implement Dolorous Edd (#602)
* Allow actions to be activated from hand.

Certain actions, such as those for Dolorous Edd and Ser Alliser Thorne,
are activated while they are in the player's hand instead of while they
are in the play area. These can now be implemented by specifying a
`location` property on the action, indicating where the action may be
activated from. The `location` property defaults to the play area.

* Implement Dolorous Edd.