Ensure persistent plot effects apply only once. (#619)
* Ensure persistent plot effects apply only once.

Previously, there were scenarios in which a plot with persistent effects
could end up being applied to the same card twice. Wraiths in their
Midst, for instance, would get applied twice when it was revealed as the
second plot during a round. Now the effect engine checks whether the
target is already in the current targets list for an effect before
adding it again.

Fixes #615.

* Fix test name.