Add card targeting to ability API (#629)
* Add `target` property to ability API.

Implements a targeting API. When a card requires a target (uses the word
'choose'), it is now possible to specify the target selection prompt
using the `target` property. If no valid targets can be chosen, then the
ability won't trigger. Once a target is chosen, it will be placed on
`context.target` for the handler. If there are multiple targets
required, then the `targets` property can be used, where each subkey
specifies a separate target.

* Convert a few cards to use the target API.