Fix effects when condition changes on enter play. (#709)
When a condition changed from another card entering play (e.g. A Song of
Summer applying its effect, then a Winter plot being revealed), the
effect was no longer being cancelled. This was due to the change in the
reapply method as the condition was already false when reaching it. Now
reapply looks at the condition whether or not the result changed from
the last check.