Give reduction to player that activated Kingsroad. (#747)
Previously, the Kingsroad was always giving its reduction to the player
set in the `controller` property. However, because the Kingsroad was
sacrificed prior to applying the reduction, the controller was always
the original owner. This lead to a bug where the Kingsroad would apply
its reduction improperly to the owner even when it had been taken
control using cards like Euron Crow's Eye.

Fixes #569.