"When Revealed" should fire per card not player. (#774)
Previously, the onPlotRevealed event used to trigger "When Revealed"
abilities was only passing the player object. This made it necessary to
set controller to null on Varys's Riddle and Pulling the Strings because
they fire the onPlotRevealed event to replicate the opponent plot. This
would lead to them triggering themselves and prompting the player to
choose order of resolution. However, by setting the controller to null,
any cards that modified the current plot (including simple ones like the
+1 gold from Roseroad) would start to produce errors.

By sending the plot card that is firing with the onPlotRevealed event,
it is now possible to trigger only that specific card's "When Revealed"
ability and is no longer necessary to disown cards that replicate other
plots.

Fixes #769.