Implement simultaneous kill event (#811)
* Convert onCharacterKilled to new event style.

* Implement simultaneous event windows.

Adds a simultaneous event window that allows an event to be applied to
multiple cards at once. An outer event is fired for a list of cards,
then an inner event is fired for each card in the list. The new window
properly maintains the division of interrupts and reactions for each
card in contrast with multiple windows being opened by simply looping
through the cards and firing individual events.

* Add simultaneous `killCharacters` method.

* Convert cards to use simultaneous kills.