Simplify prompt buttons representing cards. (#813)
Previously, when a menu prompt button was associated with a card, there
was a lot of repetition in that the text, argument and card summary were
almost always the same. Now, it is possible to just specify the card and
the other arguments are automatically filled in. If the `text` or `arg`
properties are specified, they will override the default card-generated
values.

Additionally, instead of sending the full card summary, only the
information used for card previews is sent to the client.