Add support for "up to X stat" selection prompts. (#814)
Previously, cards that had a mechanism to select cards up to a combined
strength or cost (e.g. King Robert's Warhammer) were manually tracking
the sum by monitoring card selection being toggled. Now the select card
prompt supports this natively, allowing a `maxStat` function to be
passed to indicate what the maximum stat value is, and a `cardStat`
function to take a card and get the appropriate stat value for it (e.g.
strength).