Properly implement "when marshalled" abilities. (#816)
Previously, "when marshalled" abilities were checking to see if the
current phase was marshal. This was not correct, as there are
potentially card effects that can put a card into play during the
marshal phase which should not be considered marshalling the card.

Now, the `playingType` parameter is checked to make sure that the card
has been marshalled. The `onCardEntersPlay` event has been converted to
the new event style.

This fixes bugs with Yoren, Newly Made Lord, and Melisandre where the
phase / playing type wasn't being checked at all.