Fix selectable card state when selecting opponent cards. (#824)
* Pass player object directly into serializers

Previously, a boolean flag was passed into each serializer to specify
that the object should return data as if the active player was viewing
the object. This was done to hide information from the opponent and
spectators. Now, the full player object is passed in.

* Track player card selection on the player object.

Fixes being able to see cards that are selectable that you control
(while still preventing the opponent / specators from seeing what you
can select) and simplifies marking which cards you have selected.