Fix unblanking conditional dynamic effects. (#870)
Previously, blanked effects continued to track the targets they were
applied to. However, it did not check the condition for the effect when
it was unblanked. This lead to crashes when unblanked while the
condition was no longer valid and the effect was dynamically calculated.
For example, it would try to apply Jhogo's strength bonus outside a
challenge if blanked during a challenge, leading to a crash.

Now, blanking is handled similar to conditional effects in that all
targets are removed when blanked, then targets are added back from
scratch when unblanked.