Prevent being decked in test checking winner. (#880)
The until end of phase vs at end of phase test is explicitly checking
the winner, but did not have enough cards in the deck to prevent the
draw deck from becoming empty. When it became empty, the wrong winner
was being declared. Now the test has enough cards in it to prevent from
being decked, they're just not used in the test for any other purpose.