Don't recalculate challenge strength after win. (#883)
The Haunted Forest's bonus strength was being unapplied after the
challenge winner was determined but before unopposed power was granted.
Thus when strengths were checked for unopposed power, the defender
strength was 0 and power was granted. To prevent this, challenge
strength is no longer recalculated after a winner has been determined,
and the unopposed check looks at winner and loser strength instead of
attacker and defender.