Combine windows during challenge initiation (#1162)
* Implement onChallengeInitiated event
* This event combines and replaces onChallenge, onAttackersDeclared and onBypassedByStealth

* Encapsulate event checks in ability windows.

Previously, usage of the event associated with an ability window was
scattered in multiple places. Now all usages and checks happen inside of
the ability window classes, making it easier to change internally.

* Add support for atomic events.

Adds an 'atomic event' that fires events for a series of events
atomically in order to put any interrupts / reactions for those
individual events within the same prompt.

* Combine challenge initiation events.

The events for challenge initiated, attackers declared, and individual
stealth bypasses are now fired within the same window.

* Encapsulate stealth data on challenge.

* Extract base class for ability windows.

* Fix case-sensitive requires.