Fix fetch script (promo cards) (#2220)
* Fix fetch script (promo cards)

This quick fix will grab normal card images when there is a promo. The problem was that the API was returning two rows for each card with a promo, listing the promo first. The promo did not have valid image url.

Once there are images associated with promo cards this will need to be reviewed and fixed again

* Fix linting

* spacing