Pin base image node to major version 10
The version of ZMQ we depend on has an incompatibility with the latest
version of node (13). This change pins us to a version that is
compatible with our deps and consistent with production.

Side note - using something like `node:10-alpine` will also not work
as is because the stripped down image is missing other tools (e.g.
python) the node-gyp modules need to build.