Use explicit inspect port
This is an admitedly pedadantic change. Using `--inspect` will launch a
process that (by default) binds to `127.0.0.1:9229 and waits for a
debugging client to connect. Swapping to a different IP should keep the
default port but that didn't feel obvious to me so I'm suggesting we
make the port explicit.