Docker-Compose Updates (#982)
* Pin base image node to major version 10

The version of ZMQ we depend on has an incompatibility with the latest
version of node (13). This change pins us to a version that is
compatible with our deps and consistent with production.

Side note - using something like `node:10-alpine` will also not work
as is because the stripped down image is missing other tools (e.g.
python) the node-gyp modules need to build.

* Use different image names rather than versions

When a service has both `build` and `image` attributes, after build that
image will get tagged with whatever is in `image`. Since the lobby and
node services are distinct services it might be nice to have different
image names for them instead of differentiating by version.

* Use explicit inspect port

This is an admitedly pedadantic change. Using `--inspect` will launch a
process that (by default) binds to `127.0.0.1:9229 and waits for a
debugging client to connect. Swapping to a different IP should keep the
default port but that didn't feel obvious to me so I'm suggesting we
make the port explicit.

* Expose mongo port

This will allow folks to explore keyteki data with mongodb clients
running on their host machines.

* Update docker-compose.yml