Fix to Thorium Plasmate
MoveOnBattlelineAction was determine a new index for moved cards based
on a battlelien with the card being moved already in it.

The "move" happen by first splicing out the card to be moved then
splicing it back it. This means we should calculate our "slice back in"
index assuming the moved card is not actually in the battleline.

Fixes #939