WoE: implement Yxl the Iron Captain
This is the first time multiple creatures are capturing an amount that
is more than 1, so it required a change to make sure too much amber
wasn't taken from the opponent.  For now the leftover just goes to the
last-clicked creature, but maybe we want to improve that interface a
bit in the future.

Issue: #2903