GR: `placeUnder` should remove from pile before `onLeavesPlay`
If we call `onLeavesPlay` first, the controller of the card will
revert back to the owner, and the owner will be the one to remove it
from the corresponding pile (even if the card is actually currently
controlled by the opponent).  The result is that the card will remain
in the original pile and also be a child card of the new parent, which
causes chaos.

Closes: #3825