GR: implement 3LL-I0T (#3642)
This card is the first card that forces a specific parent for an
upgrade on play (as opposed to re-attaching an existing upgrade).  I
was able to make this work by allowing a parent to be passed into
`PlayUpgradeAction` from a new `GameAction` but it's not my favorite
code in the world and I'm happy to change it however anyone else
suggests.