GR: make Be Our Geist work for treachery and enters play abilities
`treachery`/`entersPlayUnderOpponentControl` was always using the
card's owner's opponent as the new controller, but that is not correct
in cases like Be Our Geist or Lateral Shift, where you're playing
someone else's creature as if it were in your hand.  Instead, use the
active player's opponent (if they would otherwise be taking control of
the put-into-played creature).

Similarly, "entersPlay" effects like `entersPlayReady`, which have a
player-dependent condition (such as the tide being high, or being
haunted) need to calculate their effects taking into account the
active player / new controller of the card. This requires updating the
game state to recalculate all the new conditions.

Closes: #3863