GR: cards cannot self-purge from a hidden zone
The current ruling for Missile Officer Myers's scrap ability, when
discarding Multi-Dimensional Rescue, is that MDR can return itself to
your hand before it purges itself, and therefore cannot purge itself
because it is now in a hidden zone.

This would be true for other actions that purge themselves as well,
like Library Access.

I'm not sure why we were using `moveCard` for those purges, when the
`purge()` existed, so use that in all relevant places as well.

Closes: #3861