GR: fix crash when Legerdemain is discarded by Myers
For some reason, the `gameAction` in Legerdemain's `then:` ability is
fully evaluated when it's play effect is triggered by Missile Officer
Myers's scrap ability.  When there are no creatures to target, this
leads to a crash trying to resolve `preThenContext.target.controller`.
This doesn't happen when it's played normally and there are no
creatures to target though, so something else weird is going on but
this is just a quick fix for now.