Fix shuffling on ReturnToDeckAction
If the action targets 0 cards, and the opponent happened to have an
empty discard pile, `ReturnToDeckAction` would incorrectly shuffle the
opponent's deck rather than the active player's deck.

Instead, fix it to explicitly specify the player whose deck is
supposed to be shuffled, if it's not the active player's deck.  Do
this in all the cards that shuffle the opponent's deck.

This also fixes a few cards that are supposed to shuffle both player's
decks, but wouldn't if no cards from a player were being shuffled back
in.  Also fixes a bug in Mating Season that was awarding amber by
control, rather than by ownership.

Closes: #3960