GR: fix Tick-Tock message to not log cards archived from hand (#3935)
Tick-Tock can archive cards either from discard or hand, and we should
not reveal the cards archived from hand. This gets a bit messy since,
by the time the message are logged, we can't tell where the cards were
originally.  So we have to remember the player's hand and use that to
construct the message.

Closes: #3932