AS: fix Doppelganger infinite loop (#4298)
2 Doppelgangers next to each other can keep choosing each other, and
keep giving each other a new copy of their text box, leading to an
infinite loop that can only be broken by the choice of the player,
under Keyforge's Infinite Loops clause.  So in this situation, give
the user the optional to break out of the loop by making copying a
neighbor optional only if one of those neighbors is also a
shapeshifter.

Unfortunately, `properties.optional` is not something that could
previously be evaluated dynamically, so that needed to be added.