Blank effects to extend to all effects related to text box
Blanking did not previously apply to keywords granted to cards, or
effects granted to cards by an unblanked card.  For example, when an
upgrade is attached to a blanked creature that grants it a keyword, it
should not do anything.

There was a Shadow of Dis test that explicitly was testing for the
opposite of what it should be doing.

It is a bit brittle to specify the difference between effects that
relate to the text box of a creature, and those that do not (which
would not be blanked by other effects), ubt I do my best here with an
explicit list of those that should be unblankable.

Closes: #4308