Fix the default house for Redemption cards (#4327)
If a card from a different house in an old set became a Redemption
card in ToC, the `HouseId` for that card (either Redemption for ToC,
or the old house for old decks) will be explicitly marked for
newly-imported decks.  However for old decks it is still blank, but
since Redemption is in the newest pack data it will default to
Redemption incorrectly.

Instead, explicitly list default houses for these old cards.