The Promised Blade should only trigger once each turn as creature
When The Promised Blade is a creature, and the players have an equal
number of creatures (including The Promised Blade), it could trigger
twice at the start of the round if it moved, because the condition on
the second reaction became true after it moved.  Instead, make sure it
only triggers once each round.

Closes: #4374