server: fix imports for decks that contain "Annihilation Ritual"
Looks like someone changed the card ID for the card "Annihilation
Ritual".  It used to contain an extra space at the end of the name,
leading to an ID of "annihilation-ritual-" in the DB.  Now the cards
don't have that extra space, making them unable to match against the
old cards.

An alternate fix might be to re-fetch all the card data for CotA, and
update all the decks in the DB to use the new name.