PV: first attempt at prophecies and fates
Each deck is going to have 4 prophecy cards.  IRL this will be two
double-sided cards.  For now, in the deck data, we can represent this
as an array of 4 cards, and we ensure only one of each pair is active
at any time.

During Step 3, a player may "activate" one of their unactivated
prophecies by placing one card from their hand underneath it.  Only
one prophecy may be activated per turn.

The prophecy describes a condition that can be "fulfilled" on your
opponent's turn.  If it is, and the prophecy is active, your opponent
resolves the "Fate" ability of the card under the prophecy, if there
is one.  That card then goes to the discard pile and the fate
deactivates.

This is just a first draft, as official rules aren't even available
yet.