server: allow ordering a make token action even if deck is empty (#4523)
For some actions, their targets can change as a result of another
simultaneous action.  Allow that to happen specifically for making a
token creature with an empty deck.  There may be other actions that
need this in the future, but the general solution is too onerous (it
would require ordering assault vs hazardous for every fight, even when
those values are 0).