Guards against empty deck in Book of leQ
Fixes #4687

The `propertyFactory` passed to `ConditionalAction` cases
needs to be able to execute regardless of which condition
will be ultimately chosen.

In the case of Book of leQ, we need to handle evaluating the
properties of the “true” branch (change house to the top
card of the deck’s house) even when there’s no top card of
the deck.