Fixes a typo of “activeplayer” in HousePhase
Code was `activeplayer` but should be `activePlayer`.

With the bug, if The Archivist was in the archives and Memrox the Red
was in play, no cards would leave the archives because in all cases
`card.owner !== undefined`.

This adds a regression test to ensure correct interaction between those
two cards. It moves the `clickPrompt('logos')` out of `beforeEach` and
into the tests so that the new test can archive Shooler before the archive
prompt appears. (As well as choose Mars.) It also changes Ancient Bear
to Blypyp to keep the house count at 3.