Changes when swap boolean is set for rematch (#4693)
Fixes #4666

Because the 2nd player has to okay the rematch, this waits to
actually modify the `game.swap` field until that confirmation
is made.

Prevents a bug where the `swap` boolean was modified before
the 2nd player confirmed, leading to incorrect decks if they
declined the swap but then offered their own rematch.