Removes targetLocation from cardLastingEffect properties
Because CardLastingEffectAction always sets `match` to an
actual card rather than a function, passing a `targetLocation` value
into its `CardEffect` is meaningless because `CardEffect#getTargets`
(the only `Effect` code that uses `targetLocation`) is never called.

All that providing a `targetLocation` to `lastingCardEffect` did was disable
the check in `canAffect` that made `CardLastingEffectAction` only apply to
in-play cards. This check was irrespective of `targetLocation`’s actual value.

This commit removes the `targetLocation`property in `CardLastingEffectAction`
and replaces it with `allowedLocations`. This has a default of `["play area"]` and an
allowed special value of `"all"`. It updates `canAffect` to use `allowedLocations`
properly.

Since the old behavior treated all `targetLocation` values the same (i.e. “don’t
require the target to be in `"play area"`), this conversion errors on the side of
preserving the existing behavior and replaces pretty much every use of
`targetLocation` with `allowedLocations: "all"`. (There are subtle bugs that can
come up if it instead introduced new restrictions; for example, Mimicry had a
`targetLocation` of `"hand"`, but actually it needs to work no matter where it’s
played from, such as the `"deck"` with Wild Wormhole.

This also fixes #4371, a bug where a Came Back Wrong that fizzled on play
(perhaps due to re-killing its target creature) would become an upgrade in
the discard pile and stay that way even as it was shuffled back and re-drawn. In its
case (and similar effects like Poltergeistoids), we want the default behavior
of only working when it is actually in the play area, rather than the broken
behavior that it ironically got with `targetLocation: "play area"` that caused it
to be applied even if it was already discarded.

Fixes #4371