Fixes effect duration for tokens that change type/attributes (#4702)
* Fixes RemoveAllTokensAction’s target

`RemoveAllTokensAction` (used by Flying Broomstick) creates
`RemoveTokenAction`s correctly based on the selected card’s
token counts, but then applied the actions to `context.source`,
which was the triggering card, instead of the targeted card.

For example, in the case of Flying Broomstick, `context.source`
was the creature that had the Flying Broomstick upgrade and
reaped, not the creature that was targeted by the effect.

This changes the resolving to use the targeted card as its target.

Updates the Flying Broomstick tests to check stun (which is
a counter that Flying Broomstick should remove) to validate this
case, since the æmber removal and damage healing are handled by other
game actions that weren’t buggy.

Also changes to set damage directly before triggering Flying Broomstick,
for clarity.

* Fixes token effects that change other cards

Fixes #4688

This addresses bugs in how Vial of Mutation, Revo Hooligans, and
Eclectic Ambrosius (Evil Twin) implemented their effects to add
the “mutant” trait, switch to house Mars, or blank the text box,
respectively.

The previous implementations for the first two of those cards did not
consistently remove their effects when the tokens were removed, since
they listened to `onAddToken` instead of `onRemoveToken`. Additionally,
they did not correctly implement that losing the token but then
regaining it should restore the effect, based on the FAQ entry for
Sneklifter’s similarly lasting effect.

Eclectic Ambrosius (Evil Twin) incorrectly applied the text box–blanking
effect to any creature that got an ignorance counter, not just ones
targeted by the “After Reap:” effect. This would have incorrectly
blanked Envoy of Ekwirrĕ’s text box if that card swapped the ignorance
counter on to itself.

To make it more convenient to use the `condition` callback for
`cardLastingEffect`, passes the event as a new second argument. (The alternative
would be to pass a `function` and access `this`, though that’s more fragile
long-term.)

This also keeps the type condition from the previous, slightly-incorrect Eclectic
Ambrosius behavior and adds it to the other cards, so that their effects, which
mention card type specifically, deactivate when the card no longer has that type.

Explicitly adds a test of that to Eclectic Ambrosius.

---------

Co-authored-by: Jeremy Stribling <strib@alum.mit.edu>