Removes effects before registering handlers when moving cards
Fixes #3483

This solves issue #3483 for Nirbor Flamewing, where when it hit the discard
as a token creature, its onTurnStart reaction to resurrect itself wouldn’t be
registered.

By swapping the order of `updateEffects` and `updateAbilityEvents`, we
allow for the `CopyCard` effect to be expired first via `updateEffects`, so
that when `updateAbilityEvents` runs it sees the card’s actual set of
reactions, not the stale ones from the copied card.

Adds regression test to Nirbor Flamewing.