Adds a direction to FlipAction for Gĕzdrutyŏ
Based on a rules discussion on the Sanctumonious Discord,
we concluded that Gĕzdrutyŏ the Arcane’s ability, because it
specifically says “flip facedown,” should not cause a token
creature that has copied that ability to flip face-up.

This adds a `direction` property to the action that can be checked
to ensure that the flip only happens in one direction.