“Re-flips” cards before discarding them in FlipAction
Fixes #3957

This changes FlipAction so that non-creature cards are reset to their
printed type before they’re discarded. This keeps them from being
seen as creatures during `Card#onLeavesPlay`, so their æmber is
returned to the common supply rather than going to the opponent’s pool.

Adds a test to De-Animator to ensure that its existing behavior is
preserved, namely that just because a card’s printed type is creature,
if it’s destroyed as an artifact then æmber on it goes to the
common supply.